Constrain width/height of divs and images to the right/bottom of the page.
Specifically also tweak the scaling of images to reflect their reduced size.
Note that this doesn't constrain left/top to be a minimum of zero. The left
side is working fine by itself for whatever reason. The top side, for the
calendars I'm working with, has the spiral binding and doesn't get printed
anyway. Might need fixing for photo books, not tested.